src/requirements.txt
django-ninja==1.1.0
djangorestframework-simplejwt==5.3.1